+++ packages/plugin-flip/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-flip
+
+flip an image.
