+++ packages/plugin-flip/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@jimp/plugin-flip",
+  "version": "0.3.5",
+  "description": "flip an image.",
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "npm run build:node:production && npm run build:module",
+    "build:watch": "npm run build:node:debug -- -- --watch --verbose",
+    "build:debug": "npm run build:node:debug",
+    "build:module": "cross-env BABEL_ENV=module babel src -d es --source-maps --config-file ../../babel.config.js",
+    "build:node": "babel src -d dist --source-maps --config-file ../../babel.config.js",
+    "build:node:debug": "cross-env BABEL_ENV=development npm run build:node",
+    "build:node:production": "cross-env BABEL_ENV=production npm run build:node"
+  },
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "@jimp/utils": "^0.3.4"
+  },
+  "peerDependencies": {
+    "jimp": "0.3.5"
+  }
+}
